Reset moved, retain changed
Because the attack was always retained, if you missed a reset message, the attackinprogress will always remain.
Also the ResetAttackState() was initiated after the async connect, but because the connection was not setup yet, the status would not reset. So I moved it to the connection method. I've created a seperate method for setting status, you can now just use a boolean to set or reset.
Now the status will reset when you reboot the ESP.